Moved SecurityConfiguration to security package
